WIP: removed one return value from FindTeam
After much context loading we removed the `found` return value from
FindTeam from the concourse.Client struct. We brought back the
switch..case statement so we can return a custom error message when we
get a 404. All other errors are returned as is. In the fly code we can
now return the error from go-concourse as-is and it should be a helpful
error message ("team 'doesnotexist' not found").

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>