Change StatusCode to 403 for GetTeam
We were looking over the behaviour of this particular endpoint while
writing some other tests for fly. We realized that it was weird to
return a 401 when the user is already authenticated. When looking at
other API endpoints that do similar checkes to acc.IsAuthorized() they
return 403's instead of 401's. We think this was a mistake due to them
using "unauthorized" in their debug message.

With fly, when 401 is returned the user is prompted to login. This
doesn't make sense as logging in wouldn't solve the users problem;
they're already logged in! Returning 403 allows fly to give a more
sensible message which you'll see in a later commit!

Signed-off-by: Izabela Gomes <igomes@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>